[vendor-assigner] Cleaning up unused test cases (#49)
After PR#48 was merged, these TCs became outdated and unused. Since they were flaky from the start, this commit is to remove them and their unused functions.
There was also an update on a package that is needed, and after this is merged each person needs to run a `npm install` to complete the update.