[vendor-assigner] Refactoring Change Statuses test (#51)
Main focus of this commit is the refactor of the change partner assigned to a job status. I went into the same direction as the other refactor which was to split the steps and validation in each of the tests.

There was also some small fixes on the addingPartner tests that were missing and the removal of some now unused functions.